Make applyDelta be boring and fast
If I go this route, the functions that "parse" deltas into
data structures or apply them to chunks will have to go.

The problem with parsing deltas and applying them to chunks is
all the hard-to-measure memory overhead, not to mention time
overhead (several seconds).  I'm going to try to achieve the
same memory savings with polysplit and interning.